Jump while attacking Update
You can now cancel any attack that's being casted by jumping. This customizable through code. You can now manually prevent this (instead of being forced to have it as a feature) but checking if the character is attacking or charging, and they are, it's just as simple as not calling the 'Jump' code